+++ src/admin/Default/enable_game_processing.php
@@ -4,6 +4,10 @@ $game = SmrGame::getGame(Smr\Request::getInt('game_id'));
 $game->setEnabled(true);
 $game->save(); // because next page queries database
 
+// Create the Newbie Help Alliance
+require_once(get_file_loc('nha.inc.php'));
+createNHA($game->getGameID());
+
 $msg = '<span class="green">SUCCESS: </span>Enabled game ' . $game->getDisplayName();
 
 Page::create('skeleton.php', 'enable_game.php',
